% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.cc.R
\name{add.cc}
\alias{add.cc}
\title{Add carrying capacity to stage using distributions of parameters derived from Matthews et al.}
\usage{
add.cc(stage, load = F, model = "power", name.out = "new", export = F)
}
\arguments{
\item{export}{}
}
\description{
Add carrying capacity to stage using distributions of parameters derived from Matthews et al.
}
