% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_functions.R
\name{add_ground_truth}
\alias{add_ground_truth}
\title{Add ground truth inferential targets to parameter grid}
\usage{
add_ground_truth(parameter_grid, get_ground_truth)
}
\arguments{
\item{parameter_grid}{The parameter grid}

\item{get_ground_truth}{A function that inputs a subset of the parameters and
outputs an object containing the ground truth inferential target(s).}
}
\value{
An updated parameter grid
}
\description{
Add ground truth inferential targets to parameter grid
}
