% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.models.R
\name{apply.models}
\alias{apply.models}
\title{Apply models}
\usage{
apply.models(s, p0, pop.var.models, t0, t1, export = F, name.out = "new")
}
\arguments{
\item{s}{a stage object.}

\item{p0}{a populations object}

\item{pop.var.models}{a pop.var.models object}

\item{t0}{a numeric specifying the time of the previous interval.}

\item{t1}{a numeric specifying the time of the current interval.}

\item{export}{If TRUE, stage will be saved as an Rds file. Default is FALSE.}

\item{name.out}{A string to be included in the Rds output file name if export = TRUE. Default is "new".}
}
\value{
A populations object.
}
\description{
Applies models described in pop.var.models object and returns an updated populations object.
}
\examples{
# make stage
s <- make.stage(n.col = 5, n.row = 5, ar = 400)

# add a random variable to stage
s <- add.var.stage(s, var = matrix(runif(25, min = 10, max = 50), 5, 5), var.name = "SV1")

# define population variable seeds
PVs <- set.pop.var.seeds(min.via.pop = "off", new.var.name = "PV1", new.var.seed = function() runif(1,10,50))

# define population models
PVm <- set.pop.var.models(pop.var.seeds = PVs, new.var.model = list("PV1" = "fixed"))

# generate seed populations
p0 <- gen.seed.pops(stage = s, pop.var.seeds = PVs, n = 10)

# run function
p1 <- apply.models(s = s, p0 = p0, pop.var.models = PVm, t0 = 100, t1 = 90)
}
