% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functs.R
\name{assign_vars_to_global_env}
\alias{assign_vars_to_global_env}
\title{Assign variables to global environment}
\usage{
assign_vars_to_global_env(simulatr_spec, function_name, var_values)
}
\arguments{
\item{simulatr_spec}{a \code{simulatr_specifier} object}

\item{function_name}{the name of a function in the object; either \code{generate_data_function} for the data generation function, or the name of a specific method function.}

\item{var_values}{the ordered list of arguments passed to the function.}
}
\value{
NULL; has the side effect of assigning the values in the list to the named arguments in the global environment.
}
\description{
This function is meant to be used in conjunction with \code{check_simulatr_specifier_object}. The latter function detects
errors in a \code{simulatr_specifier} object. When an error occurs, \code{check_simulatr_specifier_object} returns the list of arguments
passed to a function in \code{simulatr_specifier}, along with an informative error message. The function \code{assign_vars_to_global_env}
takes a list of arguments, a \code{simulatr_specifier} object, and the name of a function in the \code{simulatr_specifier} object, and assigns
the named variables with values initialized from the list of arguments to the global environment.
}
