% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.stage.R
\name{check.stage}
\alias{check.stage}
\title{Check custom stage object adheres to required structure}
\usage{
check.stage(input, name.out = "new", export = F)
}
\arguments{
\item{input}{A list with 3 elements (regions, dimensions, distances).}

\item{name.out}{A string to be included in the Rds output file name if export = TRUE. Default is "new".}

\item{export}{If TRUE, stage will be saved as an Rds file. Default is FALSE.}
}
\value{
A basic stage object.
}
\description{
check.stage() checks that a custom stage object adheres to the basic structure required by simulatR and assigns
the "stage" class if it passes all the checks. The function checks that the first three elements of the object,
"regions", "dimensions", and "distances", are all matrices. It then checks that "dimensions" and "distances" are comprised
entirely of numeric values. It then checks that there are the same number of cells in the "regions" and "dimensions" elements.
Finally, it checks that the "distances" object includes values for every possible combination of regions.
}
\examples{
# create a mock stage object
s <- list("regions" = matrix(seq(1,9,1), 3, 3), "dimensions" = matrix(400, 3, 3),
"distances" = matrix(20, 9, 9))

# check object and assign "stage" class if it conforms.
stage <- check.stage(s)

}
\references{
Arrhenius, O. (1921). Species and area. \emph{Journal of Ecology}, 9, 95-99.
}
