% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functs.R
\name{check_simulatr_specifier_object}
\alias{check_simulatr_specifier_object}
\title{Check simulatr specifier object}
\usage{
check_simulatr_specifier_object(simulatr_spec, B_in = NULL)
}
\arguments{
\item{simulatr_spec}{a simulatr_specifier object}

\item{B_in}{number of resamples to use}
}
\value{
if no errors, the list of results; if errors occur, the ordered list of arguments passed to the function in which the error occurred.
}
\description{
Checks a simulatr specifier object for errors or warnings.
}
\details{
Note: possibly add timer and parallelization.
}
\examples{
simulatr_spec <- readRDS("/Users/timbarry/research_offsite/glmeiv/private/simulations/sim_spec_1.rds")
check <- check_simulatr_specifier_object(simulatr_spec, 2)
}
