% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_script.R
\name{collate_result_list}
\alias{collate_result_list}
\title{Collate result list}
\usage{
collate_result_list(result_df, proc_id, row_idx, method)
}
\arguments{
\item{result_df}{data frame outputted by method function}

\item{proc_id}{the processor id}

\item{row_idx}{the row of the parameter grid}

\item{method}{string giving the method}
}
\value{
a collated data frame
}
\description{
Collate result list
}
