% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functs.R
\name{compute_bias}
\alias{compute_bias}
\title{Compute bias}
\usage{
compute_bias(tbl, key, simulatr_specifier)
}
\arguments{
\item{tbl}{data frame with columns target and value. Target should have entry "estimate."}

\item{key}{data frame with columns for parameter and grid_row_id}

\item{simulatr_specifier}{a simulatr specifier object}
}
\value{
a 1-row tibble with columns value, lower_mc_ci, and upper_mc_ci
}
\description{
Computes the bias of an estimator.
}
