% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functs.R
\name{compute_coverage}
\alias{compute_coverage}
\title{Compute coverage}
\usage{
compute_coverage(tbl, key, simulatr_specifier)
}
\arguments{
\item{tbl}{data frame with columns target, value, and id. Column "target" should have entries "confint_lower" and "confint_higher"}

\item{key}{data frame with columns for parameter and grid_row_id}

\item{simulatr_specifier}{a simulatr specifier object}
}
\value{
a 1-row tibble with columns value, lower_mc_ci, and upper_mc_ci
}
\description{
Compute coverage
}
