% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functs.R
\name{compute_mse}
\alias{compute_mse}
\title{Compute mean squared error}
\usage{
compute_mse(tbl, key, sim_spec)
}
\arguments{
\item{tbl}{similar to bias}

\item{key}{similar to bias}

\item{sim_spec}{similar to bias}
}
\value{
similar to bias
}
\description{
Computes the MSE of an estimator.
}
