% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functs.R
\name{compute_rejection_probability}
\alias{compute_rejection_probability}
\title{Computes rejection probability}
\usage{
compute_rejection_probability(tbl, threshold)
}
\arguments{
\item{tbl}{data frame with columns target, value, and id. Column "target" should have entry "p_value."}

\item{threshold}{rejection threshold}
}
\value{
a 1-row tibble with columns value, lower_mc_ci, and upper_mc_ci
}
\description{
Rejection probability is type-I error (under null) and type-II error (under alternative)
}
