% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_batch_effect.R
\name{constant_batch_effect}
\alias{constant_batch_effect}
\title{Add constant batch effect to the
simulated data.}
\usage{
constant_batch_effect(n, p, bias_func_args)
}
\arguments{
\item{n}{The number of samples}

\item{p}{The number of features (e.g. genes)}

\item{bias_func_args}{The arguments given by the user
for the batch effect function}
}
\value{
A matrix of size n*p
}
\description{
\code{constant_batch_effect()} returns a dataset of
size n*p with b number of batches, p number of affected features affected by
given value s.
}
\details{
n and p define the dimension of the matrix to return.
b defines the batches the samples belong to. f defines the number of
features will be affected. s defines how much each batch is effected.
}
\examples{
\dontrun{constant_batch_effect(n=10,p=10,bias_func_args(
list(b=c(1,1,1,2,2,2,3,3,4,4),f=4,s=c(0,1,1,2))))}
\dontrun{constant_batch_effect(n=5,p=5,bias_func_args(b=2,f=4,s=c(0,1)))}
\dontrun{constant_batch_effect(n=5,p=5,bias_func_args(b=2,f=4,s=1))}
}
