% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_functions.R
\name{create_param_grid_fractional_factorial}
\alias{create_param_grid_fractional_factorial}
\title{Create param grid}
\usage{
create_param_grid_fractional_factorial(varying_values, baseline_values)
}
\arguments{
\item{varying_values}{a named list giving the values over which to vary the variables}

\item{baseline_values}{the baseline values of the variables; the variables take this value when not varying}
}
\value{
a parameter grid
}
\description{
Creates a parameter grid. For each variable, varies the value of that variable while holding the other variables constant at a specified value.
}
\examples{
varying_values <- list(x = seq(-2,2,0.1),
y = seq(0,2,0.1),
z = c("cond1", "cond2"))
baseline_values <- list(x = 0, y = 0.5, z = "cond1")
param_df <- create_param_grid_fractional_factorial(varying_values, baseline_values)
}
