% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_functions.R
\name{create_param_grid_two_way_factorial}
\alias{create_param_grid_two_way_factorial}
\title{Create param grid (two way factorial)}
\usage{
create_param_grid_two_way_factorial(param_vals, arm_names, arm_param)
}
\arguments{
\item{param_vals}{a named list of length two, giving the values to vary over}

\item{arm_names}{the names of the arms}

\item{arm_param}{the parameter to use to assign the arm names}
}
\value{
a parameter grid
}
\description{
Create param grid (two way factorial)
}
\examples{
param_vals <- list(g_perturbation = seq(0.5, 6, 0.5), pi = seq(0.1, 0.3, 0.1))
arm_names <- c("pi_small", "pi_intermediate", "pi_big")
arm_param <- "pi"
create_param_grid_two_way_factorial(param_vals, arm_names, arm_param)
}
