% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_utilities.R
\name{createSeed}
\alias{createSeed}
\title{Create Seed}
\usage{
createSeed(numberOfValues = 1, minValue = 1e+06, maxValue = 9999999, ...)
}
\arguments{
\item{numberOfValues}{a single integer value. Number of seeds to create, default is \code{1}.}

\item{minValue}{a single integer value. The minimum value that a seed can have, default is \code{1000000}.}

\item{maxValue}{a single integer value. The maximum value that a seed can have, default is \code{9999999}.}

\item{...}{optional arguments.}
}
\value{
an integer value or vector containing one or more seeds.
}
\description{
Returns one or more true random numbers which can be used as seed.
}
\details{
RANDOM.ORG offers true random numbers to anyone on the Internet.
The randomness comes from atmospheric noise, which for many purposes
is better than the pseudo-random number algorithms typically used
in computer programs. For more information see \url{https://www.random.org}.
}
\seealso{
\link{getSimulatedTwoArmMeans}
}
