% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.R
\name{dispersal}
\alias{dispersal}
\title{Generate new populations through dispersal}
\usage{
dispersal(s, p0, t0, t1, export = F, name.out = "new")
}
\arguments{
\item{s}{a stage object.}

\item{p0}{a populations object.}

\item{t0}{a numeric specifying the time of the previous interval.}

\item{t1}{a numeric specifying the time of the current interval.}

\item{export}{If TRUE, stage will be saved as an Rds file. Default is FALSE.}

\item{name.out}{A string to be included in the Rds output file name if export = TRUE. Default is "new".}
}
\value{
A populations object.
}
\description{
Apply default model of dispersal to population object and generate new object. Default configurations can factor in carrying capacity or ignore it.
}
\examples{
# make stage
s <- make.stage(n.col = 5, n.row = 5, ar = 400)

# add carrying capacity to stage
s <- add.cc(s)

# define population variable seeds
PVs <- set.pop.var.seeds(min.via.pop = "off")

# generate seed populations
p0 <- gen.seed.pops(stage = s, pop.var.seeds = PVs, n = 10)

# run function
p1 <- dispersal(s = s, p0 = p0, t0 = 100, t1 = 90)
}
