% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_noise.R
\name{exponential_noise}
\alias{exponential_noise}
\title{Create exponential noise}
\usage{
exponential_noise(n, p, noise_func_args)
}
\arguments{
\item{n}{The number of samples}

\item{p}{The number of features (e.g. genes)}

\item{noise_func_args}{Takes the function arguments}
}
\value{
A matrix of size n*p
}
\description{
\code{exponential_noise()} returns a dataset of size n*p from a
exponential distribution
}
\details{
n and p define the dimension of the matrix to return.
noise_func_args defines the argument for generating distribution
(in this case uniform). We can define vector of rates as
noise_func_args$rate=1.
}
\examples{
\dontrun{exponential_noise(10,10,list(rate=1))}
}
