% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functs.R
\name{fdp}
\alias{fdp}
\title{Evaluate the FDP of a discovery set}
\usage{
fdp(output, ground_truth)
}
\arguments{
\item{output}{The output of a method, an object that must contain the field nonnulls}

\item{ground_truth}{The ground truth, an object that must contain the field nonnulls}
}
\value{
The FDP
}
\description{
Evaluate the FDP of a discovery set
}
