% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.diag.R
\name{get.diag}
\alias{get.diag}
\title{Get diagonally aligned cells for each cell in a matrix}
\usage{
get.diag(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A list of length(mat) where each element (i) contains the index vectors of each cell diagonally offset from mat[i].
}
\description{
Get diagonally aligned cells for each cell in a matrix
}
\examples{
m = matrix(1,3,3)
diagonals <- get.diag(m)
}
