% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Extract dataset}
\usage{
get_dataset(gse)
}
\arguments{
\item{gse}{The GSE number}
}
\value{
A matrix extracted from GEOquery package
}
\description{
\code{simulate_gse()} returns a dataset for a given GSE number.
}
\details{
gse is a GEO accession number. This can be found in NCBI website.
A Series record defines a set of related Samples considered to be
part of a group, how they are ordered.A Series provides a focal
point and description of the experiment as a whole. Series records
may also contain tables describing extracted data, summary conclusions,
or analyses. Each Series record is assigned a unique and stable GEO
accession number (GSExxx).
}
\examples{
\dontrun{
get_dataset(gse = "GSE3821")
}
}
