% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_constructors.R
\name{get_example_simulatr_specifier}
\alias{get_example_simulatr_specifier}
\title{get example simulatr_specifier}
\usage{
get_example_simulatr_specifier()
}
\value{
an example simulatr specifier object
}
\description{
get example simulatr_specifier
}
\examples{
load_all()
get_example_simulatr_specifier()
}
