% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gpl_data.R
\name{get_gpl_data}
\alias{get_gpl_data}
\title{Extract data}
\usage{
get_gpl_data(gplnum)
}
\arguments{
\item{gplnum}{The GPL number}
}
\value{
A matrix extracted from GEOquery package
}
\description{
\code{get_gpl_data()} returns a information for a given GPL number.
}
\details{
gple is a GEO accession number. This can be found in NCBI website.
A Platform record describes the list of elements on the array (e.g., cDNAs,
oligonucleotide probesets, ORFs, antibodies) or the list of elements that
may be detected and quantified in that experiment (e.g., SAGE tags,
peptides). Each Platform record is assigned a unique and stable GEO
accession number (GPLxxx). A Platform may reference many Samples that
have been submitted by multiple submitters.
}
\examples{
\dontrun{
get_dataset("GPL96")
}
}
