% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functs.R
\name{get_memory_used}
\alias{get_memory_used}
\title{Compute memory used since last call to gc(reset = TRUE)}
\usage{
get_memory_used()
}
\value{
Number of bytes used
}
\description{
Compute memory used since last call to gc(reset = TRUE)
}
