% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_script.R
\name{get_ordered_args}
\alias{get_ordered_args}
\title{Get ordered arguments for function object}
\usage{
get_ordered_args(function_object, simulatr_spec, row_idx)
}
\arguments{
\item{function_object}{The function object}

\item{simulatr_spec}{Simulatr specifier object}

\item{row_idx}{The row of the parameter grid from which to get parameter values}
}
\value{
List of ordered arguments
}
\description{
Get ordered arguments for function object
}
