% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_simluate_continuous_data.R
\name{getSimluatedTwoArmMeans}
\alias{getSimluatedTwoArmMeans}
\title{Get Simluated Two Arm Means}
\usage{
getSimluatedTwoArmMeans(
  n1,
  n2,
  mean1,
  mean2,
  sd1,
  sd2,
  ...,
  seed = NA_integer_,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{n1}{a single integer value. The sample size of group 1.}

\item{n2}{a single integer value. The sample size of group 2.}

\item{mean1}{a single numeric value. The assumed mean of group 1.}

\item{mean2}{a single numeric value. The assumed mean of group 2.}

\item{sd1}{a single numeric value. The assumed standard deviation of group 1.}

\item{sd2}{a single numeric value. The assumed standard deviation of group 2.}

\item{...}{ensures that all arguments (starting from the "...") are to be named.}

\item{seed}{a single integer value. The start value for generating pseudo-random numbers.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided" (default), "greater" or "less".}
}
\description{
Simluates and returns a normally distributed continuous dataset for two groups.
}
\details{
If the \code{seed} is not specified (NA) the function \link{createSeed} will be used to create
a true random number as seed.

See \url{https://cran.r-project.org/package=simstudy} for a comprehensive R package for study data simulation.
}
\examples{
getSimluatedTwoArmMeans(n1 = 50, n2 = 50, mean1 = 5, mean2 = 7, sd1 = 3, sd2 = 4, seed = 123)

}
\seealso{
\link{createSeed}
}
