% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_simulate_continuous_data.R
\name{getSimulatedTwoArmMeans}
\alias{getSimulatedTwoArmMeans}
\title{Get Simulated Two Arm Means}
\usage{
getSimulatedTwoArmMeans(n1, n2, mean1, mean2, sd1, sd2, seed = NA_integer_)
}
\arguments{
\item{n1}{a single integer value. The sample size of group 1.
TODO document all arguments}
}
\value{
a list with all arguments and results; the output is defined as a
class with name 'SimulationResult'.
}
\description{
Simulates and returns a normally distributed continuous dataset for two
groups.
}
\details{
TODO describe
}
\examples{
getSimulatedTwoArmMeans(n1 = 50, n2 = 50, mean1 = 5, mean2 = 7, sd1 = 3,
                        sd2 = 4, seed = 123)

}
