% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit.avg.disp.dist.R
\name{inherit.avg.disp.dist}
\alias{inherit.avg.disp.dist}
\title{Average dispersal distance inheritance model}
\usage{
inherit.avg.disp.dist(p0, p, t0, t1, min = 0.005, max = NULL, SF = 1)
}
\arguments{
\item{p0}{populations object from previous time bin.}

\item{p}{the numeric index of the population in p0$population.variables.}

\item{t0}{time of previous time bin.}

\item{t1}{time of current time bin.}

\item{SF}{a numeric value by which the changes in ADD at speciation event can be scaled. For flexibility.}
}
\value{
a numeric value.
}
\description{
Changes to ADD that occur during speciation event. Tied to abundance of population (large population = broad distribution new ADD can be sampled from.
}
\keyword{"PV1",}
\keyword{#}
\keyword{0.1,}
\keyword{1),}
\keyword{10)}
\keyword{100,}
\keyword{400)}
\keyword{5,}
\keyword{90)}
\keyword{<-}
\keyword{=}
\keyword{PVs}
\keyword{PVs,}
\keyword{Run}
\keyword{a}
\keyword{ar}
\keyword{define}
\keyword{function}
\keyword{function()}
\keyword{gen.seed.pops(stage}
\keyword{generate}
\keyword{inherit.avg.disp.dist(p0}
\keyword{internal}
\keyword{make}
\keyword{make.stage(n.col}
\keyword{n}
\keyword{n.row}
\keyword{new.var.name}
\keyword{new.var.seed}
\keyword{p}
\keyword{p0}
\keyword{p0,}
\keyword{pop.var.seeds}
\keyword{population}
\keyword{populations}
\keyword{runif(1,10,50))}
\keyword{s}
\keyword{s,}
\keyword{sample(1:length(p0$population.variables),}
\keyword{seed}
\keyword{set.pop.var.seeds(pop.gen.threshold}
\keyword{stage}
\keyword{t0}
\keyword{t1}
\keyword{variables}
