% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_corr_matrix.R
\name{is_corr_matrix}
\alias{is_corr_matrix}
\title{Tests if object is a correlation matrix}
\usage{
is_corr_matrix(x, type = c("full", "triangular"))
}
\arguments{
\item{x}{square matrix with real values}

\item{type}{"full" or "triangular". See details.}
}
\value{
boolean
}
\description{
Tests if an object is a valid correlation matrix, ie if the input object is
symmetric positive semidefinite matrix with unit diagonal.
}
\details{
The function will terminate with an error if the input object is a not a
square matrix with numeric values. For convenience, the input matrix can also
be given in lower/upper triangular form by specifying
\code{type = "triangular"}.
}
\examples{
x <- diag(c(1, 1, 1))
is_corr_matrix(x)
x[upper.tri(x)] <- 0.5
is_corr_matrix(x)
is_corr_matrix(x, type = "triangular")
}
