% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_triangular.R
\name{is_triangular}
\alias{is_triangular}
\title{Check if triangular matrix}
\usage{
is_triangular(x, type = c("any", "lower", "upper"))
}
\arguments{
\item{x}{square matrix}

\item{type}{String. Type of triangular matrix that is being tested.
See details.}
}
\value{
boolean
}
\description{
Checks if square matrix is a triangular matrix.
}
\details{
A triangular matrix can be either a lower triangular matrix or an upper
triangular matrix. By default (\code{type = "any"}) the function tests
if the square matrix is a either lower or upper triangular. If
\code{type = "lower"} or \code{type = "upper"} is specified,
testing if the matrix is specifically lower or upper triangular is
possible.
}
\examples{
x <- matrix(c(1, 0, 1, 1), nrow = 2, ncol = 2)
is_triangular(x)
}
