% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_platforms.R
\name{list_platforms}
\alias{list_platforms}
\title{List all platform numbers}
\usage{
list_platforms()
}
\value{
list of all platforms along with information about them
}
\description{
\code{lists_platforms()} returns a list of platforms that can be
found in NCBI website
}
\details{
Provides information about all the platforms
}
\examples{
\dontrun{
list_platforms()
}

}
