% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.stage.R
\name{make.stage}
\alias{make.stage}
\title{Create basic stage object}
\usage{
make.stage(
  n.col,
  n.row,
  ar,
  export = F,
  name.out = "new",
  distance.alt = F,
  od.path = "average"
)
}
\arguments{
\item{n.col}{The number of columns desired for this stage.}

\item{n.row}{The number of rows desired for this stage.}

\item{ar}{The area of each region. Can be a single numeric value, in which case all regions are assigned the same area, or a numeric vector of length n.col*n.row, in which case area values are assigned by index.}

\item{export}{If TRUE, stage will be saved as an Rds file. Default is FALSE.}

\item{name.out}{A string to be included in the Rds output file name if export = TRUE. Default is "new".}

\item{distance.alt}{Either FALSE (default) or an alternative function for calculating the distances between regions.}

\item{od.path}{How to calculate the distance between regions which are not diagonally, longitudinally, or latitudinally aligned with one another when the default distance calculation method is used.
For such distances, the default method calculates two hypotenuses using Pythagoras' theorem; one where the adjacent side is latitudinally aligned with the origin region, and one where
it is aligned with the destination region. "min" assigns the smallest of these two hypotenuses as the distance between these regions. "max" the largest".
:average" (the default) assigns the mean value.}
}
\value{
A basic stage object.
}
\description{
'make.stage()' generates an object of class "stage", which is a list with at least three elements (more can be added with other functions). The first is a matrix with each cell (hereafter region) assigned an identifying number (the 'regions').
The second is a corresponding matrix presenting the area values assigned to each of these regions (the 'dimensions'). The third is a distance matrix where each column and row represents a region,
and the values within the distances separating them (the 'distances').
}
\details{
The default distance calculation method assumes each region to be a square and takes the square root of the area
as the latitudinal and longitudinal dimensions of the region. Pythagoras' theorem is applied to these latitudinal and longitudinal distances to get the distance to cross these regions diagonally.
Distances between regions are calculated assuming that species have to migrate from the center of one region to the center of another. As such, when the distance separating two
regions is calculated, half the latitudinal, longitudinal, or diagonal distance (depending on direction of travel) of the origin and destination regions is added to the distance values of any regions
the species have to traverse. The distances separating regions that are not longitudinally, latitudinally, or diagonally aligned are also calculated using Pythagoras' theorem, using the summed latitudinal
distances of all regions comprising the adjacent side and the summed longitudinal distances of all regions comprising the opposite. Two hypotenuses are calculated; one where the adjacent side is aligned
with the origin region, and one where it is aligned with the destination. The minimum, maximum, or average of these distances are then assigned as the distance between the two regions, depending on 'od.path'.

A separate distance calculation method can be provided via the argument 'distance.alt'. This should be provided in the form of a function with arguments 'stage' and 'dimensions', which accept matrices formatted
as described.

Outputted as an object by default. Can be exported as an Rds file.
}
\examples{
# all regions have the same area
stage <- make.stage(3,3,100)

# regions have different areas
stage <- make.stage(3,3,c(100,200,300,400,500,600,700,800,900))

}
