% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.abundance.R
\name{model.abundance}
\alias{model.abundance}
\title{Default abundance model}
\usage{
model.abundance(
  a0,
  min.via.pop = F,
  MVP.value,
  pop.gen = F,
  g.diff,
  pg.threshold = 0.1,
  t0,
  t1,
  SF = 1
)
}
\arguments{
\item{a0}{a numeric value specifying abundance in previous time bin.}

\item{min.via.pop}{if TRUE, populations with a0 below MVP.value will have a greater likelihood of declining (the further below MVP.value, the greater the likelihood). Default is FALSE.}

\item{MVP.value}{the minimum viable population number used if min.via.pop = TRUE.}

\item{pop.gen}{if TRUE, populations with g.diff > pg.threshold will tend to decrease.}

\item{g.diff}{A numeric value. Should be change in population across previous time interval.}

\item{pg.threshold}{A numeric value specifying the threshold above which change in the mean genome is considered detrimental to the survival of the population. Default is 0.1.}

\item{t0}{time of previous time bin.}

\item{t1}{time of current time bin.}

\item{SF}{a numeric value by which the changes in abundance can be scaled.}
}
\value{
a numeric value.
}
\description{
Changes are drawn from normal distributions where the standard deviation is the product of the time passed, and a factor (SF). If min.via.pop is TRUE
and the abundance of the population is below MVP.value, then the mean of this normal distribution shifts by the difference.
}
\keyword{internal}
