% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_corr_matrix.R
\name{nearest_corr_matrix}
\alias{nearest_corr_matrix}
\title{Nearest correlation matrix}
\usage{
nearest_corr_matrix(
  x,
  type = c("full", "triangular"),
  verbose = FALSE,
  max_ite = 100L,
  conv_cri = 1e-08
)
}
\arguments{
\item{x}{a matrix}

\item{type}{"full" or "triangular". See details.}

\item{verbose}{boolean. Should info for each iteration be printed to the log?}

\item{max_ite}{positive integer. Maximal number of iteration used to try to
find nearest correlation matrix.}

\item{conv_cri}{positive numeric value. Convergence criteria applied when
comparing infinite norms to determine convergence of algorithm.}
}
\value{
matrix
}
\description{
Finds the nearest correlation matrix given a symmetric matrix with numeric
values.
}
\details{
Finding the nearest correlation matrix is done by iteratively alternating
between projecting \code{x} onto the set of positive semidefinite matrices
and the set of matrices with unit diagonal. This iterative process is
guarantied to converge to a matrix with unit diagonal that is also positive
semidefinite, ie a correlation matrix. See references for further details.

Note that the return object should be tested to assert whether or not a
nearest correlation matrix was actually found, since it is not a guaranty
that the default maximum number of iterations is sufficient in all cases.

For convenience, the input matrix can be given in lower/upper triangular form
by specifying \code{type = "triangular"}.
}
\examples{
x <- matrix(c(1, 0.3, 0.9, 0.3, 1, 0.9, 0.9, 0.9, 1), nrow = 3, ncol = 3)
nearest_corr_matrix(x)
}
\references{
Higham, N.J. (1988). Computing a nearest symmetric positive semidefinite matrix.

Higham. N.J. (2002). Computing the nearest correlation matrix - a problem from finance.

Wicklin, R. (2013). Simulating data with SAS.
}
