% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_data.R
\name{normal_data}
\alias{normal_data}
\title{Create normal data
Gaussian or normal distribution}
\usage{
normal_data(n, p)
}
\arguments{
\item{n}{The number of samples}

\item{p}{The number of features (e.g. genes)}
}
\value{
A matrix of size n*p
}
\description{
Create normal data
Gaussian or normal distribution
}
\examples{
\dontrun{
normal_data(10, 10)
}

}
