% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.model.R
\name{parse.model}
\alias{parse.model}
\title{Parse and evaluate a model}
\usage{
parse.model(m, s, p0, t0, t1, p, r)
}
\arguments{
\item{p0}{A populations object created with gen.seed.pops or simulate.}

\item{t0}{A numeric value specifying previous time bin.}

\item{t1}{A numeric value specifying current time bin.}

\item{p}{A string specifying the name of the population.}

\item{r}{A numeric value specifying the region the process is occurring in/to.}

\item{model}{A model object created with specify.model.}

\item{stage}{A stage object created with make.stage or checked with check.stage.}
}
\value{
A numeric value.
}
\description{
Works on models of all levels.
}
\examples{
# make stage
s <- make.stage(n.col = 5, n.row = 5, ar = 400)

# add a random variable to stage
s <- add.var.stage(s, var = matrix(runif(25, min = 10, max = 50), 5, 5), var.name = "SV1")

# define population variables
PVs <- set.pop.var.seeds(min.via.pop = "off", new.var.name = "PV1", new.var.seed = function() runif(1,10,50))

# generate seed populations
p0 <- gen.seed.pops(stage = s, pop.var.seeds = PVs, n = 10)

# First example: evaluating a binary level-1 model
# define model
m1 <- specify.model(s = s, p = p0, type = "binary", variables = c("SV1", "PV1"), expression = "PV1 <= SV1", ID = "m1")

# parse and evaluate model - population and region chosen randomly
parse.model(m = m1, s = s, p0 = p0, t0 = 100, t1 = 90, p = sample(1:length(p0$population.variables), 1), r = sample(s$regions, 1))

# Second example: evaluating a continuous level-1 model
m2 <- specify.model(s = s, p = p0, type = "continuous", variables = c("SV1", "PV1"), expression = "abs(PV1 - SV1)", ID = "m2")

# parse and evaluate model
parse.model(m = m2, s = s, p0 = p0, t0 = 100, t1 = 90, p = sample(1:length(p0$population.variables), 1), r = sample(s$regions, 1))
}
\keyword{internal}
