% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.nested.model.R
\name{parse.nested.model}
\alias{parse.nested.model}
\title{Parse and evaluate a nested model}
\usage{
parse.nested.model(m, s, p0, p, r)
}
\arguments{
\item{p0}{A populations object created with gen.seed.pops or simulate.}

\item{p}{A string specifying the name of the population.}

\item{r}{A numeric value specifying the region the process is occurring in/to.}

\item{model}{A nested model object created with specify.nested.model.}

\item{stage}{A stage object created with make.stage or checked with check.stage.}
}
\value{
A numeric value.
}
\description{
Works on models of all levels.
}
\examples{
# make stage
s <- make.stage(n.col = 5, n.row = 5, ar = 400)

# add a random variable to stage
s <- add.var.stage(s, var = matrix(runif(25, min = 10, max = 50), 5, 5), var.name = "SV1")

# define population variables
PVs <- set.pop.var.seeds(min.via.pop = "off", new.var.name = "PV1", new.var.seed = function() runif(1,10,50))

# generate seed populations
p0 <- gen.seed.pops(stage = s, pop.var.seeds = PVs, n = 10)

# First example: evaluating a binary level-2 model
# First, create level-1 continuous model
m1 <- specify.model(s = s, p = p0, type = "continuous", variables = c("SV1", "PV1"), expression = "abs(PV1 - SV1)")

# Next, create a level-2 binary model
m2 <- specify.nested.model(s = s, p = p0, m = m1, type = "binary", variables = c("PV1", "m"), expression = "1 <= PV1/m")

parse and evaluate model
parse.model(m = m2, s = s, p0 = p0, p = sample(p0$population.IDs, 1), r = sample(s$regions, 1))
}
\keyword{internal}
