% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_add_corr_vars.R
\name{plan_add_corr_vars}
\alias{plan_add_corr_vars}
\title{Simulate correlated variables}
\usage{
plan_add_corr_vars(x, var_def, corr_matrix = NULL, plan_label = NULL)
}
\arguments{
\item{x}{simulatr_plan object}

\item{var_def}{todo}

\item{corr_matrix}{todo}

\item{plan_label}{todo}
}
\value{
a simulatr_plan object
}
\description{
Simulate variables with with provided correlations. Correlation are only
approximately if one or more discrete variables are simulated.
}
\details{
TODO:
\enumerate{
\item Figure out where the inspiration for this approach comes from, so that
proper references can be provided. Does not seem like Austin did this in any
of his propensity score simulation studies?
\item Expand details on induced correlations, eg that we are using pearsons
correlation coefficient. Look a the math and figure out why correlations
are preserved when transforming to other continuous distirbutions, but
correlations are watered down when transforming to discrete distributions.
Wgat happens if other correlation coefficients are used, eg spearman?
}
}
\examples{
2 + 2
}
