% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_simluate_continuous_data.R
\name{plot.SimulationResult}
\alias{plot.SimulationResult}
\title{Plot Simulation Result}
\usage{
\method{plot}{SimulationResult}(
  x,
  ...,
  main = "Continuous Fake Data",
  xlab = "Group",
  ylab = "Simulated Values"
)
}
\arguments{
\item{x}{a \code{SimulationResult} object to plot.}

\item{...}{ensures that all arguments (starting from the "...") are to be named.}

\item{main}{an overall title for the plot.}

\item{xlab}{a title for the x axis.}

\item{ylab}{a title for the y axis.}
}
\value{
A ggplot2 object.
}
\description{
Generic function to plot a \code{SimulationResult} object.
}
\details{
Uses ggplot2 to create the plot.
}
\examples{
x <- getSimluatedTwoArmMeans(n1 = 50, n2 = 50, mean1 = 5, mean2 = 7, sd1 = 3, sd2 = 4, seed = 123)
if (require(ggplot2)) plot(x)

}
