% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functs.R
\name{plot_all_arms}
\alias{plot_all_arms}
\title{Plot all arms}
\usage{
plot_all_arms(summarized_results, parameter, metric)
}
\arguments{
\item{summarized_results}{a data frame outputted by \code{summarize_results.}}

\item{parameter}{name of the target parameter}

\item{metric}{name of the metric used to quantify inference on the target parameter}
}
\value{
a plot
}
\description{
Plots all arms of a simulation study for a given parameter and metric.
}
