% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_noise.R
\name{poisson_noise}
\alias{poisson_noise}
\title{Create poisson noise}
\usage{
poisson_noise(n, p, noise_func_args)
}
\arguments{
\item{n}{The number of samples}

\item{p}{The number of features (e.g. genes)}

\item{noise_func_args}{Takes the function arguments}
}
\value{
A matrix of size n*p
}
\description{
\code{poisson_noise()} returns a dataset of size n*p from a
Poisson distribution
}
\details{
n and p define the dimension of the matrix to return.
noise_func_args defines the argument for generating distribution
(in this case uniform). We can define vector of non-negative means
as noise_func_args$lambda=2.
}
\examples{
\dontrun{
poisson_noise(10, 10, list(lambda = 1))
}
}
