% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functs.R
\name{power}
\alias{power}
\title{Evaluate the power of a discovery set}
\usage{
power(output, ground_truth)
}
\arguments{
\item{output}{The output of a method, an object that must contain the field nonnulls}

\item{ground_truth}{The ground truth, an object that must contain the field nonnulls}
}
\value{
The power
}
\description{
Evaluate the power of a discovery set
}
