% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_simluate_continuous_data.R
\name{print.SimulationResult}
\alias{print.SimulationResult}
\title{Print Simulation Result}
\usage{
\method{print}{SimulationResult}(x, ...)
}
\arguments{
\item{x}{a \code{SimulationResult} object to print.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function to print a \code{SimulationResult} object.
}
