% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.simulatr_plan.R
\name{print.simulatr_plan}
\alias{print.simulatr_plan}
\title{Title}
\usage{
\method{print}{simulatr_plan}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{eh}
}
\value{
print
}
\description{
Title
}
\examples{
2 + 2
}
