% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_psd.R
\name{project_psd}
\alias{project_psd}
\title{Project real symmetric matrix onto the set of positive semidefinite matrices.}
\usage{
project_psd(x)
}
\arguments{
\item{x}{a real symmetric matrix}
}
\value{
a matrix
}
\description{
Projects a real symmetric matrix onto the set of positive semidefinite
matrices, by using eigendecomposition, ie X = QDQ*, where Q is the matrix of
eigenvectors and D is the diagonal matrix containing the eigenvalues. The
projection is done by replacing any negative eigenvalues in D by zero, and
then reassembling the matrix.
}
\examples{
x <- diag(c(-1, 0))
project_psd(x)
}
