% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_noise.R
\name{random_noise}
\alias{random_noise}
\title{Create random noise}
\usage{
random_noise(n, p, noise_func_args)
}
\arguments{
\item{n}{The number of samples}

\item{p}{The number of features (e.g. genes)}

\item{noise_func_args}{Takes the function arguments}
}
\value{
A matrix of size n*p
}
\description{
\code{random_noise()} returns a dataset of size n*p from a
gaussian distribution
}
\details{
n and p define the dimension of the matrix to return.
noise_func_args defines the argument for generating distribution
(in this case gaussian/normal). We can define standard deviation
of the function as noise_func_args$sd=0.5
}
\examples{
\dontrun{random_noise(10,10,list(sd=1))}
}
