% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAR_power.R
\name{SAR_power}
\alias{SAR_power}
\title{Calculate carrying capacity for each region using power model with default parameters}
\usage{
SAR_power(area)
}
\arguments{
\item{area}{A matrix specifying the size (in hectares) of each region of a stage.}
}
\value{
A matrix specifying the species carrying capacity of each region of a stage.
}
\description{
SAR_power() uses the power model (Arrhenius, 1921) to estimate the species carrying capacity of each region of a stage.
For each region, parameters c and z are drawn from Weibull distributions with shape and scale values derived from the datasets of
Matthews et al. (in prep).
}
\examples{
# create a mock area object
example <- matrix(20,4,4)

# estimate species carrying capacity for each region
capacities <- SAR_power(example)
}
\references{
Arrhenius, O. (1921). Species and area. \emph{Journal of Ecology}, 9, 95-99.
}
