% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.abundance.R
\name{seed.abundance}
\alias{seed.abundance}
\title{Default method of assigning starting abundance values to populations}
\usage{
seed.abundance(minimum = 10)
}
\arguments{
\item{minimum}{a numeric value specifying the minimum acceptable abundance number. Default is 10.}
}
\value{
a numeric value.
}
\description{
Values are drawn from a Weibull distribution fitted to the data of Traill et al. Minimum value returned is 10 (the minimum estimate derived from the data of Traill et al.).
}
\examples{
# Run the function
seed.abundance()
}
\references{
Traill, L. W., Bradshaw, C. J. A., & Brook, B. W. (2007). Minimum viable population size: A meta-analysis of 30 years of published estimates. \emph{Biological Conservation}, 139, 156-166.
}
\keyword{internal}
