% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.avg.disp.dist.R
\name{seed.avg.disp.dist}
\alias{seed.avg.disp.dist}
\title{Default method of assigning starting avg.disp.dist values to populations}
\usage{
seed.avg.disp.dist(cutoff = 0.006)
}
\arguments{
\item{cutoff}{a numeric value specifying the largest avg.disp.dist value that will be re-sampled. Default is 0.006 (minimum of datasets).}
}
\value{
a numeric value.
}
\description{
Values are drawn from a gamma distribution fitted to the data of Kinlan & Gaines (2003).
}
\examples{
# Run the function
seed.avg.disp.dist()
}
\references{
Kinlan, B. P. & Gaines, S. D. (2003). Propagule dispersal in marine and terrestrial environments: a community perspective. \emph{Ecology}, 84, 2007-2020.
}
\keyword{internal}
