% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.disp.prop.R
\name{seed.disp.prop}
\alias{seed.disp.prop}
\title{Default method of assigning dispersal propensity values to populations}
\usage{
seed.disp.prop(cutoff = 0)
}
\arguments{
\item{cutoff}{a numeric value specifying the largest avg.disp.prop value that will be re-sampled. Default is 0.}
}
\value{
a numeric value.
}
\description{
Values are drawn from a uniform distribution bounded between 0 and 1. Only produces nonzero values by default.
}
\examples{
# Run the function
seed.disp.prop()
}
\keyword{internal}
