% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.gen.het.R
\name{seed.gen.het}
\alias{seed.gen.het}
\title{Default method of assigning starting genetic heterogeneity values to populations}
\usage{
seed.gen.het(cutoff = 0)
}
\arguments{
\item{cutoff}{a numeric value specifying the largest gen.het value that will be re-sampled. Default is 0.}
}
\value{
a numeric value.
}
\description{
Values are drawn from a uniform distribution bounded between 0 and 1. Only produces nonzero values by default.
}
\examples{
# Run the function
seed.gen.het()
}
