% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.min.via.pop.R
\name{seed.min.via.pop}
\alias{seed.min.via.pop}
\title{Default method of assigning minimum viable population values to populations}
\usage{
seed.min.via.pop(minimum = 10)
}
\arguments{
\item{minimum}{a numeric value specifying the minimum acceptable viable population number. Default is 10.}
}
\value{
a numeric value.
}
\description{
Values are drawn from a Weibull distribution fitted to the data of Traill et al. Minimum value returned is 10 by default (the minimum estimate derived from the data of Traill et al.).
}
\examples{
# Run the function
seed.min.via.pop()
}
\references{
Traill, L. W., Bradshaw, C. J. A., & Brook, B. W. (2007). Minimum viable population size: A meta-analysis of 30 years of published estimates. \emph{Biological Conservation}, 139, 156-166.
}
\keyword{internal}
