% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.pop.gen.threshold.R
\name{seed.pop.gen.threshold}
\alias{seed.pop.gen.threshold}
\title{Default method of assigning starting pop.gen.threshold values to populations}
\usage{
seed.pop.gen.threshold(minimum = 0.1)
}
\arguments{
\item{minimum}{a numeric value specifying the minimum acceptable pop.gen.threshold number. Default is 0.1.}
}
\value{
a numeric value.
}
\description{
Values are drawn from a uniform distribution.
}
\examples{
# Run the function
seed.pop.gen()
}
