% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed.vagility.R
\name{seed.vagility}
\alias{seed.vagility}
\title{Default method of assigning starting vagility values to populations}
\usage{
seed.vagility(cutoff = 0)
}
\arguments{
\item{cutoff}{a numeric value specifying the largest vagility value that will be re-sampled. Default is 0.}
}
\value{
a numeric value.
}
\description{
Values are drawn from a uniform distribution bounded between 0 and 1. Only produces nonzero values by default.
}
\examples{
# Run the function
seed.vagility()
}
\keyword{internal}
