% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.pop.var.inher.R
\name{set.pop.var.inher}
\alias{set.pop.var.inher}
\title{Set population inheritance models}
\usage{
set.pop.var.inher(
  pop.var.seeds,
  abundance = "fixed",
  genome = "fixed",
  avg.disp.dist = "default",
  disp.prop = "default",
  min.via.pop = "default",
  pop.gen.threshold = "fixed",
  new.var.inher = NULL,
  name.out = "new",
  export = F
)
}
\arguments{
\item{pop.var.seeds}{A pop.var.seeds object, specifying which population variables.}

\item{abundance}{Either "fixed" (the default) or a model object (built using specify.model or specify.nested.model).}

\item{genome}{Either "fixed" (the default) or a model object (built using specify.model or specify.nested.model).}

\item{avg.disp.dist}{Either "default" (the default), a model object (built using specify.model or specify.nested.model), or "fixed".}

\item{disp.prop}{Either "default" (the default), a model object (built using specify.model or specify.nested.model), or "fixed".}

\item{min.via.pop}{Either "default" (the default), a model object (built using specify.model or specify.nested.model), or "fixed".}

\item{pop.gen.threshold}{Either "fixed" (the default), "default", or a model object (built using specify.model or specify.nested.model).}

\item{new.var.inher}{A list, where each element specifies the prescribed inheritance model for each non-standard variable included. Each element of the list should share the name of the variable it pertains to. List elements should either be "fixed" or a model object (built using specify.model or specify.nested.model).}

\item{name.out}{If export = TRUE, a string that specifies the name of the output file.}

\item{export}{If TRUE, updated stage with new variable is exported as an Rds file to working directory.}
}
\value{
A pop.var.inher object.
}
\description{
set.pop.var.models generates a pop.var.inher object which specifies which population variables change at speciation events and the models that define how they do so.
Uses pop.var.seeds to determine which models to load. By default, defines models for changes in population avg.disp.dist, disp.prop, and min.via.pop, and sets all other core
population variables to fixed (i.e. do not change during speciation event). If model objects are provided, ensure they return the new numeric value for the variable at the time as their output (i.e the highest level model is continuous).
}
\examples{
# Generate pop.var.seeds object
pop.seeds <- set.pop.var.seeds()

# Generate pop.var.models object
pop.inher <- set.pop.var.inher(pop.var.seeds = pop.seeds)
}
