% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.pop.var.models.R
\name{set.pop.var.models}
\alias{set.pop.var.models}
\title{Set population models}
\usage{
set.pop.var.models(
  pop.var.seeds,
  abundance = "default",
  genome = "default",
  avg.disp.dist = "fixed",
  disp.prop = "fixed",
  min.via.pop = "fixed",
  pop.gen.threshold = "fixed",
  new.var.model = NULL,
  name.out = "new",
  export = F
)
}
\arguments{
\item{pop.var.seeds}{A pop.var.seeds object, specifying which population variables.}

\item{abundance}{Either "default" (the default), a model object (built using specify.model or specify.nested.model), or "fixed".}

\item{genome}{Either "default" (the default), a model object (built using specify.model or specify.nested.model),  or "fixed".}

\item{avg.disp.dist}{Either "fixed" (the default), or a model object (built using specify.model or specify.nested.model).}

\item{disp.prop}{Either "fixed" (the default), or a model object (built using specify.model or specify.nested.model).}

\item{min.via.pop}{Either "fixed" (the default), or a model object (built using specify.model or specify.nested.model).}

\item{pop.gen.threshold}{Either "fixed" (the default), or a model object (built using specify.model or specify.nested.model).}

\item{new.var.model}{A list, where each element specifies the prescribed model for each non-standard variable included. Each element of the list should share the name of the variable it pertains to. List elements should either be "fixed" (if variable is unchanging), a model object (built using specify.model or specify.nested.model), or a function which specifies how the new value is derived.}

\item{name.out}{If export = TRUE, a string that specifies the name of the output file.}

\item{export}{If TRUE, updated stage with new variable is exported as an Rds file to working directory.}
}
\value{
A pop.var.models object.
}
\description{
set.pop.var.models generates a pop.var.models object which specifies which population variables change over time and the models that define how they do so.
Uses pop.var.seeds to determine which models to load. By default, defines models for change in population abundance and mean genome, and sets all other core
population variables to fixed. If model objects are provided, ensure they return the new numeric value for the variable at the time as their output (i.e the highest level model is continuous).
}
\examples{
# Generate pop.var.seeds object
pop.seeds <- set.pop.var.seeds()

# Generate pop.var.models object
pop.models <- set.pop.var.models(pop.var.seeds = pop.seeds)
}
