% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_simulate_continuous_data.R
\name{showDefault.SimulationResult}
\alias{showDefault.SimulationResult}
\title{Show Simulation Result}
\usage{
showDefault.SimulationResult(x, ...)
}
\arguments{
\item{x}{a \code{SimulationResult} object to show.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function to show a \code{SimulationResult} object.
}
\examples{
getSimulatedTwoArmMeans(n1 = 50, n2 = 50, mean1 = 5, mean2 = 7, sd1 = 3, sd2 = 4, seed = 123)

}
