% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dataset.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{Simulate a Dataset}
\usage{
simulate_dataset(
  n,
  p,
  beta,
  base = NULL,
  family = "gaussian",
  cor = 0.5,
  cortype = 1,
  noise = 1,
  noisevar = 1,
  bias = 0,
  biastype = 1
)
}
\arguments{
\item{n}{The number of samples.}

\item{p}{The number of features.}

\item{beta}{The model coefficients.}

\item{base}{The dataset to use as basis for the generation.}

\item{family}{The type of model to use when generating the outcome variable.
Either \code{gaussian}, \code{binomial} or \code{cox}.}

\item{cor}{Correlation strength within the data}

\item{cortype}{Method to use for generating the correlation within the
dataset. See below for details.}

\item{noise}{Amount of noise to add to each predictor}

\item{noisevar}{Variance of noise}

\item{bias}{Amount of bias to simulate}

\item{biastype}{Method to use for generating the bias. See below for details.}
}
\value{
A matrix of size n*p.
}
\description{
\code{simulate_dataset()} returns a dataset of size n*p
}
\details{
to be done
}
\examples{
# to be done
}
