% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_glm_data.R
\name{simulate_glm_data}
\alias{simulate_glm_data}
\title{Simulate a dataframe using glm}
\usage{
simulate_glm_data(dat)
}
\arguments{
\item{dat}{The dataframe from which new data is to
be simulated}
}
\value{
A dataframe with simulated data
}
\description{
\code{simulate_glm_data()} returns a dataframe simulated using
generalized linear model
}
\examples{
\dontrun{
simulate_glm_data(base = simulatr::normal_data(5, 6))
}
\dontrun{
simulate_glm_data(dat = as.data.frame(simulatr::get_dataset(
    gse = "GSE3821"
)))
}

}
