% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gse.R
\name{simulate_gse}
\alias{simulate_gse}
\title{Simulate data from GSE}
\usage{
simulate_gse(n = 10, p = 10, gse = "GSE3821")
}
\arguments{
\item{n}{The number of samples}

\item{p}{The number of features (e.g. genes)}

\item{gse}{The GSE number}
}
\value{
A matrix of size n*p
}
\description{
\code{simulate_gse()} returns a dataset of size n*p for
a given GSE number.
}
\details{
n and p define the dimension of the matrix to return.
gse is a GEO accession number. This can be found in NCBI website.
Each Series record is assigned a unique and stable GEO
accession number (GSExxx).
}
\examples{
\dontrun{
simulate_gse(10, 10, gse = "GSE3821")
}
}
