% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{simulatr}
\alias{simulatr}
\alias{simulatr-package}
\title{simulatr - Generator for Reproducible Fake Data}
\description{
An example package for learning about software development best-practices for R packages.
}
\details{
For more information please visit \href{https://kkmann.github.io/workshop-r-swe}{kkmann.github.io/workshop-r-swe}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kkmann.github.io/workshop-r-swe}
  \item \url{https://github.com/kkmann/workshop-r-swe}
  \item Report bugs at \url{https://github.com/fpahlke/simulatr/issues}
}

}
\author{
Friedrich Pahlke, Daniel Sabanes Bove, Kevin Kunzmann
}
