% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_constructors.R
\name{simulatr_specifier}
\alias{simulatr_specifier}
\title{simulatr_specifier}
\usage{
simulatr_specifier(
  parameter_grid,
  fixed_parameters,
  generate_data_function,
  run_method_functions,
  evaluation_functions = list()
)
}
\arguments{
\item{parameter_grid}{a named data frame storing the values of study parameters that vary}

\item{fixed_parameters}{a named list storing the values of the parameters that stay fixed across parameter settings}

\item{generate_data_function}{a \code{simulatr_function} that generates the data}

\item{run_method_functions}{a named list of \code{simulatr_functions} that define the methods to be run in the simulation}

\item{evaluation_functions}{a named list of functions that define the metrics used to evaluate each method}
}
\value{
a \code{simulatr_specifier} object
}
\description{
A \code{simulatr_specifier} object defines a simulation study in the \code{simulatr} framework.
}
