% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.nested.model.R
\name{specify.nested.model}
\alias{specify.nested.model}
\title{Specify nested model}
\usage{
specify.nested.model(
  s,
  p,
  m,
  type,
  variables,
  expression,
  name.out = "new",
  export = F
)
}
\arguments{
\item{s}{A stage object}

\item{p}{A populations object}

\item{m}{Either a model object, or a list of model objects.}

\item{type}{Either "binary" or "continuous".}

\item{variables}{A character vector containing the names of the population and stage variables included in the model or NA (where there are no variables involved).}

\item{expression}{A character vector specifying the construction of the model as a single string. Model should be referred to as 'm' or, if multiple models are included, the model names or 'm1', 'm2', and so forth.}

\item{models}{A character vector containing the names of}
}
\value{
a model object
}
\description{
Creates a model object that specifies the details of a model to be used in simulation. Used to build model objects that draw on other models as variables (i.e. nested models).
Models should be nested one level at a time. For example, if you want to nest a nested model (i.e. create a level-3 model), first you need to create the level-2 nested model,
then submit that as a variable during the creation of the level-3 model. Note that if you submit a list of models as m, these will be referred to in the model by the names provided or
in the order they appear in the list. For example, if you don't assign names to your models via names(), the first model in the list will be assigned the identifier "m1", the second "m2", and so forth.
}
\examples{
# Run the function
model <- specify.model(type = "binary", variables = NA, expression = "runif(1) > 0.5")
}
