% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functs.R
\name{summarize_results}
\alias{summarize_results}
\title{Summarize results}
\usage{
summarize_results(simulatr_specifier, raw_result_df, metrics)
}
\arguments{
\item{simulatr_specifier}{a \code{simulatr_specifier} object}

\item{raw_result_df}{the data frame of raw results, as outputted by \code{simulatr}}

\item{metrics}{character vector of metrics to compute; options include coverage and bias.}
}
\value{
a data frame of summarized results
}
\description{
Summarizes the results of a \code{simulatr} simulation.
}
\examples{
\dontrun{
simulatr_specifier <-
readRDS("/Users/timbarry/research_offsite/glmeiv/private/simulations/sim_spec_1.rds")
raw_result_df <-
readRDS("/Users/timbarry/research_offsite/glmeiv/private/simulations/result.rds")
raw_result_df <- dplyr::filter(raw_result_df,
parameter \%in\% c("m_perturbation", "g_perturbation", "pi"))
metrics <- c("bias", "coverage")
}
}
