% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_triangular_corr.R
\name{trans_triangular_corr}
\alias{trans_triangular_corr}
\title{Transform triangular correlation matrix}
\usage{
trans_triangular_corr(x)
}
\arguments{
\item{x}{matrix. A correlation matrix on triangular form.}
}
\value{
matrix. Correlation matrix on full from.
}
\description{
Transforms a correlation matrix on triangular form to a correlation
matrix on full form.
}
\examples{
x <- matrix(c(1, 0, 0.5, 1), nrow = 2, ncol = 2)
trans_triangular_corr(x)
}
