% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_noise.R, R/uniform_noise.R
\name{uniform_noise}
\alias{uniform_noise}
\alias{binomial_noise}
\title{Create binomial noise}
\usage{
binomial_noise(n, p, noise_func_args)

uniform_noise(n, p, noise_func_args)
}
\arguments{
\item{n}{The number of samples}

\item{p}{The number of features (e.g. genes)}

\item{noise_func_args}{Takes the function arguments}
}
\value{
A matrix of size n*p

A matrix of size n*p
}
\description{
\code{binomial_noise()} returns a dataset of size n*p from a
binomial or Bernoulli distribution

\code{uniform_noise()} returns a dataset of size n*p from a
uniform distribution
}
\details{
n and p define the dimension of the matrix to return.
noise_func_args defines the argument for generating distribution
(in this case uniform). We can define number of trials as
noise_func_args$size=10 and probability of each trial as
noise_func_args$prob=0.5

n and p define the dimension of the matrix to return.
noise_func_args defines the argument for generating distribution
(in this case uniform). We can define minimum and maximum
of the function as noise_func_args$min=1 and noise_func_args$max=10
}
\examples{
\dontrun{binomial_noise(10,10,list(size=50,prob=0.5))}
\dontrun{uniform_noise(10,10,list(min=1,max=10))}
}
