% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_accessors.R
\name{update_B_sim_spec}
\alias{update_B_sim_spec}
\title{Update B in a simulatr_specifier object}
\usage{
update_B_sim_spec(simulatr_spec, B)
}
\value{
an updated simulatr_specifier object
}
\description{
Takes a simulatr_specifier object and a value for B; updates B and return a new object.
}
