% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.cc.R
\name{add.cc}
\alias{add.cc}
\title{Add carrying capacity to stage using distributions of parameters derived from Matthews et al. data}
\usage{
add.cc(stage, model = "power", power.non0 = F, name.out = "new", export = F)
}
\arguments{
\item{stage}{A stage object or the name of the stage Rds file as a string.}

\item{model}{Species-area model used to calculate carrying capacity. Default is power model (Arrhenius, 1921).}

\item{power.non0}{If TRUE, all carrying capacities estimated by the default power model will be 1 or higher (i.e. no zeros). Default is FALSE.}

\item{name.out}{If export = TRUE, a string that specifies the name of the output file.}

\item{export}{If TRUE, updated stage with carrying capacities is exported as an Rds file to working directory.}
}
\value{
A stage object with element "cc" added.
}
\description{
Add carrying capacity to stage using distributions of parameters derived from Matthews et al. data
}
\examples{
# create a mock stage object
s <- list("regions" = matrix(seq(1,9,1), 3, 3), "dimensions" = matrix(400, 3, 3),
"distances" = matrix(20, 9, 9))

# check object and assign "stage" class if it conforms.
stage <- check.stage(s)

# add carrying capacities
stage.cc <- add.cc(stage)
}
\references{
Arrhenius, O. (1921). Species and area. \emph{Journal of Ecology}, 9, 95-99.
}
