% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functs.R
\name{check_simulatr_specifier_object}
\alias{check_simulatr_specifier_object}
\title{Check simulatr specifier object}
\usage{
check_simulatr_specifier_object(
  simulatr_spec,
  B_in = NULL,
  return_data = FALSE,
  parallel = TRUE
)
}
\arguments{
\item{simulatr_spec}{a simulatr_specifier object}

\item{B_in}{(default determined by simulatr_spec) number of resamples to use}

\item{return_data}{(default false) whether to return the data}

\item{parallel}{(default true) parallelize execution?}
}
\value{
if no errors, the list of results; if errors occur, the ordered list of arguments passed to the function in which the error occurred.
}
\description{
Checks a simulatr specifier object for correctness.
}
\examples{
\dontrun{
simulatr_spec <-
readRDS("/Users/timbarry/research_offsite/glmeiv/private/simulations/sim_spec_1.rds")
simulatr_spec@parameter_grid  <- simulatr_spec@parameter_grid[1:10,]
simulatr_spec@fixed_parameters[["n"]] <- 2000
check <- check_simulatr_specifier_object(simulatr_spec, 5)
}
}
