% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{epi_sim}
\alias{epi_sim}
\title{Simulate an epidemic.}
\usage{
epi_sim(
  a_g = 5,
  lambda_g = 1,
  a_s = 5,
  lambda_s = 1,
  R = 2,
  psi = 0.5,
  mu = 2e-05,
  N_eff = log(100),
  init_genome = sample(c("A", "C", "G", "T"), 10000, replace = T),
  sample_dp = function(n) {
     rep(10000, n)
 },
  sample_sb = function(n) {
     rep(0, n)
 },
  min_af = 0.03,
  N = 1e+06,
  p_samp = 0.5,
  trans_samp = NA,
  coverage = 1,
  n_obs = 100,
  include_root = TRUE,
  start_date = as.Date("2000-01-01"),
  outdir = "my_epidemic",
  seed = NA
)
}
\arguments{
\item{a_g}{Shape parameter of the Gamma-distributed generation interval. Defaults to 5.}

\item{lambda_g}{Rate parameter of the Gamma-distributed generation interval. Defaults to 1.}

\item{a_s}{Shape parameter of the Gamma-distributed sojourn interval (time from inoculation to receiving a diagnostic test. Defaults to 5.}

\item{lambda_s}{Rate parameter of the Gamma-distributed sojourn interval. Defaults to 1.}

\item{R}{Basic reproductive number. Defaults to 1.5.}

\item{psi}{Second parameter of the Negative Binomial distribution.}

\item{mu}{Evolution rate, in substitutions per site per day. Defaults to 1e-6.}

\item{N_eff}{The within-host effective population size, expressed as the number of virions at time t ie exp(N_eff * t). Defaults to log(100).}

\item{init_genome}{The initial genome present in the seed of the epidemic. Defaults to a genome consisting of 1e4 random draws from A, C, G, T.}

\item{sample_dp}{A function with one argument, n, that randomly samples n read depths. Defaults to the function that always returns a constant depth of 10,000 reads.}

\item{sample_sb}{A function with one argument, n, that randomly samples n strand biases. Defaults to the function that always returns a constant strand bias of 0.}

\item{min_af}{Limit of detection for within-host variant frequencies.}

\item{N}{The size of the population. The population is assumed to be well-mixed, and to start with a single infectious individual. Defaults to 1e6.}

\item{p_samp}{The probability of being sampled. Defaults to 0.5.}

\item{trans_samp}{If not NA, a vector of length two specifying P(sampled | ancestor is sampled) and P(unsampled | ancestor is unsampled), or a vector of length 1 if these two are the same. The stationary distribution of the corresponding 2x2 transition matrix must equal (p_samp, 1 - p_samp).}

\item{coverage}{Probability that a position on the genome is sequenced. Each position is an i.i.d. draw. Defaults to 1.}

\item{n_obs}{The number of sampled cases to generate. Defaults to 100.}

\item{include_root}{Should the root (i.e. the first case to be seeded in the population) be included in the output FASTA and VCF files? Defaults to TRUE.}

\item{start_date}{Date on which the first case is inoculated. Defaults to January 1st, 2000.}

\item{outdir}{Name of the output directory. Defaults to "my_epidemic."}

\item{seed}{If integer, the seed is set to that integer. If NA, no seed is set. Defaults to NA.}
}
\value{
A directory containing a single FASTA for all simulated cases, a FASTA consisting of the reference genome, VCF files for each simulated case, and the times of sample collection relative to the inoculation time of the first case (in days).
}
\description{
This function simulates pathogen spread through a population, given user-specified parameters. It outputs genomic and epidemiological data for each case, including read-level sequencing data.
}
