% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{epi_sim}
\alias{epi_sim}
\title{Simulate an epidemic.}
\usage{
epi_sim(
  a_g = 5,
  lambda_g = 1,
  a_s = 5,
  lambda_s = 1,
  R = 1.5,
  rho = Inf,
  mu = 1e-05,
  p = 5e-06,
  v = 1000,
  lambda_b = 1.05,
  init_genome = sample(c("A", "C", "G", "T"), 10000, replace = T),
  sample_dp = function(n) {
     rep(10000, n)
 },
  sample_sb = function(n) {
     rep(0, n)
 },
  N = 1e+06,
  n_simulated = 30,
  include_root = TRUE,
  outdir = "my_epidemic"
)
}
\arguments{
\item{a_g}{Shape parameter of the Gamma-distributed generation interval. Defaults to 5.}

\item{lambda_g}{Rate parameter of the Gamma-distributed generation interval. Defaults to 1.}

\item{a_s}{Shape parameter of the Gamma-distributed sojourn interval (time from inoculation to receiving a diagnostic test. Defaults to 5.}

\item{lambda_s}{Rate parameter of the Gamma-distributed sojourn interval. Defaults to 1.}

\item{R}{Basic reproductive number. Defaults to 1.5.}

\item{rho}{Overdispersion parameter of the offspring distribution. Defaults to Inf, indicating the offspring distribution is Poisson. If finite, the offspring distribution is taken to be Negative Binomial with a mean of R.}

\item{mu}{Evolution rate, in substitutions per site per day. Defaults to 1e-6.}

\item{p}{Within-host mutation rate, in mutations per site per replication cycle. Note that mu/p gives equals number of replication cycles per day. Defaults to 5e-6.}

\item{v}{The number of offspring produced per replication cycle.}

\item{lambda_b}{The mean number of particles per transmission bottleneck. A shifted Poisson distribution is assumed, i.e. the bottleneck size is 1 + Poisson(lambda_b - 1). Defaults to 1.05.}

\item{init_genome}{The initial genome present in the seed of the epidemic. Defaults to a genome consisting of 1e4 random draws from A, C, G, T.}

\item{sample_dp}{A function with one argument, n, that randomly samples n read depths. Defaults to the function that always returns a constant depth of 10,000 reads.}

\item{sample_sb}{A function with one argument, n, that randomly samples n strand biases. Defaults to the function that always returns a constant strand bias of 0.}

\item{N}{The size of the population. The population is assumed to be well-mixed, and to start with a single infectious individual. Defaults to 1e6.}

\item{n_simulated}{The number of individuals to simulate. Defaults to 30.}

\item{include_root}{Should the root (i.e. the first case to be seeded in the population) be included in the output FASTA and VCF files? Defaults to TRUE.}

\item{outdir}{Name of the output directory. Defaults to "my_epidemic."}

\item{growth}{Exponential growth rate of cases. Defaults to NULL. Exactly one of R and growth may be specified; the other must be set to NULL.}
}
\value{
A directory containing a single FASTA for all simulated cases, a FASTA consisting of the reference genome, VCF files for each simulated case, and the times of sample collection relative to the inoculation time of the first case (in days).
}
\description{
This function simulates pathogen spread through a population, given user-specified parameters. It outputs genomic and epidemiological data for each case, including read-level sequencing data.
}
