% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.latlong.R
\name{get.latlong}
\alias{get.latlong}
\title{Get latitudinally and longitudinally aligned cells for each cell in a matrix}
\usage{
get.latlong(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A list of length(mat) where each element (i) contains the index vectors of each cell latitudinal or longitudinally offset from mat[i].
}
\description{
Get latitudinally and longitudinally aligned cells for each cell in a matrix
}
\examples{
m = matrix(1,3,3)
diagonals <- get.latlong(m)
}
