% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_accessors.R
\name{get_params_for_nextflow}
\alias{get_params_for_nextflow}
\title{Get params for nextflow}
\usage{
get_params_for_nextflow(simulatr_spec, fp)
}
\arguments{
\item{simulatr_spec}{the simulatr_specification object}

\item{fp}{a file path}
}
\description{
Given a file path fp, writes all parameters to which nextflow needs access to a plain text file stored at fp.
}
\details{
Parameters are separated by line. Each line is of the form "param:value," where "param" is the name of the parameter and "value" is its value.
}
