% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List Datasets useable as Base Datasets for Simulations}
\usage{
list_datasets(platform, source = "GEO")
}
\arguments{
\item{platform}{e.g. HG-U133B for mRNA datasets measured with the Affymetrix
HG-U133B microarray.}

\item{source}{where to search for datasets.}
}
\value{
A named list of datasets, where each dataset is represented by a list
with the following elements:
platform: platform specified as parameter
n: number of observations in the dataset
p: number of features in the dataset
download: function to actually download the dataset
}
\description{
\code{lists_datasets()} returns a list of datasets than can be passed
to \code{simulate_dataset} as basis for the simulation.
}
\details{
to be done
}
\examples{
# to be done
}
