% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R, R/list_platforms.R
\name{list_datasets}
\alias{list_datasets}
\alias{list_platforms}
\title{List Datasets useable as Base Datasets for Simulations}
\usage{
list_datasets(platform)

list_platforms()
}
\arguments{
\item{platform}{The platform number}
}
\value{
A dataframe with information about series which used a specific
platform

list of all platforms along with information about them
}
\description{
\code{lists_datasets()} returns a list of datasets that can be passed
to \code{simulate_dataset} as basis for the simulation.

\code{lists_platforms()} returns a list of platforms that can be
found in NCBI website
}
\details{
A Platform record describes the list of elements on the array (e.g.,
cDNAs, oligonucleotide, probesets, ORFs, antibodies) or the list of elements
that may be detected and quantified in that experiment (e.g., SAGE tags,
peptides). Each Platform record is assigned a unique and stable GEO accession
number (GPLxxx). A Series record defines a set of related Samples considered
to be part of a group, how the Samples are related, and if and how they are
ordered. A Series provides a focal point and description of the experiment as
a whole. Series records may also contain tables describing
extracted data, summary conclusions, or analyses. Each Series record is
assigned a unique and stable GEO accession number (GSExxx). S

Provides information about all the platforms
}
\examples{
\dontrun{list_datasets(platform = "GPL95")}

\dontrun{list_platforms()}

}
