% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.abundance.R
\name{model.abundance}
\alias{model.abundance}
\title{Default model for changes in abundance over time}
\usage{
model.abundance(A0, MVP, time, factor = 1)
}
\arguments{
\item{factor}{}
}
\description{
Default model for changes in abundance over time
}
