% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.abundance.R
\name{model.abundance}
\alias{model.abundance}
\title{Default abundance model}
\usage{
model.abundance(p0, p.ID, t.series, t0, SF = 1)
}
\arguments{
\item{p0}{populations object from previous time bin.}

\item{p.ID}{string identifying population in question.}

\item{t.series}{a numeric vector specifying time bins.}

\item{t0}{time of previous time bin.}

\item{SF}{a numeric value by which the changes in abundance can be scaled. For flexibility.}
}
\value{
a numeric value.
}
\description{
Changes are drawn from normal distributions where the standard deviation is the product of the time passed, and a factor (SF). Capable of inducing changes based on population genetics and minimum viable population values.
}
\keyword{internal}
