% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.abundance.R
\name{model.abundance}
\alias{model.abundance}
\title{Default abundance model}
\usage{
model.abundance(p0, p.ID, t.series, t0, SF = 1)
}
\arguments{
\item{p0}{populations object from previous time bin.}

\item{p.ID}{string identifying population in question.}

\item{t.series}{a numeric vector specifying time bins.}

\item{t0}{time of previous time bin.}

\item{SF}{a numeric value by which the changes in abundance can be scaled. For flexibility.}
}
\value{
a numeric value.
}
\description{
Changes are drawn from normal distributions where the standard deviation is the product of the time passed, and a factor (SF). Capable of inducing changes based on population genetics and minimum viable population values.
}
\keyword{"PV1",}
\keyword{#}
\keyword{#'}
\keyword{0.1,}
\keyword{1),}
\keyword{10)}
\keyword{100,}
\keyword{400)}
\keyword{5,}
\keyword{<-}
\keyword{=}
\keyword{Create}
\keyword{PVs}
\keyword{PVs,}
\keyword{a}
\keyword{ar}
\keyword{c(100,90,80,70,60,50,40,30,20,10)}
\keyword{define}
\keyword{function()}
\keyword{gen.seed.pops(stage}
\keyword{generate}
\keyword{internal}
\keyword{make}
\keyword{make.stage(n.col}
\keyword{model.abundance(p0}
\keyword{n}
\keyword{n.row}
\keyword{new.var.name}
\keyword{new.var.seed}
\keyword{p.ID}
\keyword{p0}
\keyword{p0,}
\keyword{pop.var.seeds}
\keyword{population}
\keyword{populations}
\keyword{runif(1,10,50))}
\keyword{s}
\keyword{s,}
\keyword{sample(1:length(p0$population.variables),}
\keyword{seed}
\keyword{set.pop.var.seeds(pop.gen.threshold}
\keyword{specifying}
\keyword{stage}
\keyword{t.series}
\keyword{t.series)}
\keyword{t0}
\keyword{times}
\keyword{variables}
\keyword{vector}
