% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.genome.R
\name{model.genome}
\alias{model.genome}
\title{Default genome model}
\usage{
model.genome(p0, p, t0, t1, SF = 1)
}
\arguments{
\item{p0}{populations object from previous time bin.}

\item{p}{the numeric index of the population in p0$population.variables.}

\item{t0}{time of previous time bin.}

\item{t1}{time of current time bin.}

\item{SF}{a numeric value by which the changes in mean genome can be scaled. For flexibility.}
}
\value{
a numeric value.
}
\description{
Changes are drawn from normal distributions where the standard deviation is the product of the time passed, and a factor (SF).
Capable of inducing changes based on principles of population genetics and abundance.
}
\keyword{"PV1",}
\keyword{#}
\keyword{0.1,}
\keyword{1),}
\keyword{10)}
\keyword{100,}
\keyword{400)}
\keyword{5,}
\keyword{90)}
\keyword{<-}
\keyword{=}
\keyword{PVs}
\keyword{PVs,}
\keyword{Run}
\keyword{a}
\keyword{ar}
\keyword{define}
\keyword{function}
\keyword{function()}
\keyword{gen.seed.pops(stage}
\keyword{generate}
\keyword{internal}
\keyword{make}
\keyword{make.stage(n.col}
\keyword{model.genome(p0}
\keyword{n}
\keyword{n.row}
\keyword{new.var.name}
\keyword{new.var.seed}
\keyword{p}
\keyword{p0}
\keyword{p0,}
\keyword{pop.var.seeds}
\keyword{population}
\keyword{populations}
\keyword{runif(1,10,50))}
\keyword{s}
\keyword{s,}
\keyword{sample(1:length(p0$population.variables),}
\keyword{seed}
\keyword{set.pop.var.seeds(pop.gen.threshold}
\keyword{stage}
\keyword{t0}
\keyword{t1}
\keyword{variables}
