% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.genome.R
\name{model.genome}
\alias{model.genome}
\title{Default genome model}
\usage{
model.genome(p0, p.ID, t.series, t0, SF = 1)
}
\arguments{
\item{p0}{populations object from previous time bin.}

\item{p.ID}{string identifying population in question.}

\item{t.series}{a numeric vector specifying time bins.}

\item{t0}{time of previous time bin.}

\item{SF}{a numeric value by which the changes in mean genome can be scaled. For flexibility.}
}
\value{
a numeric value.
}
\description{
Changes are drawn from normal distributions where the standard deviation is the product of the time passed, and a factor (SF).
Capable of inducing changes based on principles of population genetics and abundance.
}
\keyword{internal}
