% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.nested.model.R
\name{parse.nested.model}
\alias{parse.nested.model}
\title{Parse and evaluate a nested model}
\usage{
parse.nested.model(m, s, p0, t0, t1, p, r)
}
\arguments{
\item{p0}{A populations object created with gen.seed.pops or simulate.}

\item{t0}{A numeric value specifying previous time bin.}

\item{t1}{A numeric value specifying current time bin.}

\item{p}{A string specifying the name of the population.}

\item{r}{A numeric value specifying the region the process is occurring in/to.}

\item{model}{A level-2 or greater model created with specify.nested.model.}

\item{stage}{A stage object created with make.stage or checked with check.stage.}
}
\value{
A numeric value.
}
\description{
Works on models of all levels.
}
\examples{
# make stage
s <- make.stage(n.col = 5, n.row = 5, ar = 400)

# add a random variable to stage
s <- add.var.stage(s, var = matrix(runif(25, min = 10, max = 50), 5, 5), var.name = "SV1")

# define population variables
PVs <- set.pop.var.seeds(min.via.pop = "off", new.var.name = "PV1", new.var.seed = function() runif(1,10,50))

# generate seed populations
p0 <- gen.seed.pops(stage = s, pop.var.seeds = PVs, n = 10)

# Create a level-1 continuous model
m1 <- specify.model(s = s, p = p0, type = "continuous", variables = c("SV1", "PV1"), expression = "abs(PV1 - SV1)", ID = "m1")

# And then a level-1 binary model
m2 <- specify.model(s = s, p = p0, type = "binary", variables = c("SV1", "PV1"), expression = "PV1 <= SV1", ID = "m2")

# Next, create a level-2 binary model
m3 <- specify.nested.model(s = s, p = p0, m = list(m1,m2), type = "binary", variables = c("PV1"), expression = "1 <= PV1/(m1.1*m2.1)", ID = "m3")

# Try parsing
parse.nested.model(m = m3, s = s, p0 = p0, t0 = 100, t1 = 90, p = sample(1:length(p0$population.variables), 1), r = sample(s$regions, 1))

# Create another level-1 model
m4 <- specify.model(s = s, p = p0, type = "continuous", variables = c("SV1", "PV1"), expression = "PV1*SV1", ID = "m4")

# Create a level-3 continuous model
m5 <- specify.nested.model(s = s, p = p0, m = list(m3,m4), type = "continuous", variables = NA, expression = "m3.2*m4.1", ID = "m5")

# parse and evaluate model
parse.nested.model(m = m5, s = s, p0 = p0, t0 = 100, t1 = 90, p = sample(1:length(p0$population.variables), 1), r = sample(s$regions, 1))
}
\keyword{internal}
