% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_simulate_continuous_data.R
\name{print.SimulationResult}
\alias{print.SimulationResult}
\title{Print Simulation Result}
\usage{
\method{print}{SimulationResult}(x, ...)
}
\arguments{
\item{x}{a \code{SimulationResult} object to print.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic function to print a \code{SimulationResult} object.
}
\examples{
x <- getSimulatedTwoArmMeans(n1 = 50, n2 = 50, mean1 = 5, mean2 = 7, sd1 = 3, sd2 = 4, seed = 123)
print(x)

}
