% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_unit_diag.R
\name{project_unit_diag}
\alias{project_unit_diag}
\title{Project matrix onto set of matrices with unit diagonal.}
\usage{
project_unit_diag(x)
}
\arguments{
\item{x}{a square matrix}
}
\value{
matrix
}
\description{
Projects a numeric square matrix onto the set of matrices with unit diagonal,
by simply replacing the diagonal elements with ones.
}
\examples{
x <- diag(c(0, 0))
project_unit_diag(x)
}
