% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.pop.var.seeds.R
\name{set.pop.var.seeds}
\alias{set.pop.var.seeds}
\title{Set population variables}
\usage{
set.pop.var.seeds(
  avg.disp.dist = "default",
  disp.prop = "default",
  abundance = "default",
  gen.het = "default",
  min.via.pop = "default",
  stable.seeds = F,
  new.var.name = "new.variable",
  new.var.seed = NULL,
  name.out = "new",
  export = F
)
}
\arguments{
\item{avg.disp.dist}{Either "default" (the default), or a function which specifies how the initial value is derived for a population.}

\item{disp.prop}{Either "default" (the default), or a function which specifies how the initial value is derived for a population.}

\item{abundance}{Either "default" (the default) or a function which specifies how the initial value is derived for a population.}

\item{gen.het}{Either "default" (the default), "off", or a function which specifies how the initial value is derived for a population.}

\item{min.via.pop}{Either "default" (the default), "off", or a function which specifies how the initial value is derived for a population.}

\item{stable.seeds}{If TRUE, all seed abundance values will be greater than the minimum viable population numbers. Default is FALSE.}

\item{new.var.name}{If new.var.seed is a function or list of functions, a string, or list of strings, specifying the name(s) of the new population variables. Strings should not contain punctuation.}

\item{new.var.seed}{A function or list of functions specifying how the initial values for the new variables are derived for a population.}

\item{name.out}{If export = TRUE, a string that specifies the name of the output file.}

\item{export}{If TRUE, updated stage with new variable is exported as an Rds file to working directory.}
}
\value{
A pop.var.seeds object.
}
\description{
set.pop.var.seeds generates a pop.var.seeds object which specifies the population variables for a simulation and defines how their initial values will be derived. Core variables
are abundance, average dispersal distance (avg.disp.dist), and dispersal propensity (disp.prop). These variables are integral to the core processes of the simulation (speciation, dispersal, extirpation) and so
cannot be switched off. Genetic heterogeneity (gen.het) and minimum viable population (min.via.pop) are also included as optional variables. By default, abundance and minimum viable population numbers are
drawn from a Weibull distribution fitted to the data of Traill et al., with values below 10 being resampled, while the avg.disp.dist values are drawn from a gamma distribution fitted to the data of Kinlan & Gaines (2003).
By default, the gen.het and disp.prop values are derived from uniform distributions bounded between 0 and 1, with gen.het and avg.disp.dist values of 0 resampled by default. All five variables are included in the output by default.
However, min.via.pop and gen.het can be switched off, and all five variables can be replaced with custom functions if desired. This function also allows for additional custom population variables to be added as input-less functions.
}
\examples{
# Generate pop.var.seeds object
pop.seeds <- set.pop.var.seeds

}
\references{
Traill, L. W., Bradshaw, C. J. A., & Brook, B. W. (2007). Minimum viable population size: A meta-analysis of 30 years of published estimates. \emph{Biological Conservation}, 139, 156-166.

Kinlan, B. P. & Gaines, S. D. (2003). Propagule dispersal in marine and terrestrial environments: a community perspective. \emph{Ecology}, 84, 2007-2020.
}
