% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_script.R
\name{setup_script}
\alias{setup_script}
\title{Setup script}
\usage{
setup_script(simulatr_spec, B_in, function_object, row_idx)
}
\arguments{
\item{simulatr_spec}{simulation specifier object}

\item{B_in}{B to set (0 means do not update)}

\item{function_object}{either a method or data generation simulatr_function object}

\item{row_idx}{row index of the parameter grid}
}
\value{

}
\description{
Runs setup for either the generate_data.R script or run_methods.R script.
}
