% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dataset.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{Simulate a data set}
\usage{
simulate_dataset(
  n = 5,
  p = 5,
  base = simulatr::normal_data(n, p),
  beta = NULL,
  noise_func = matrix(0, n, p),
  noise_func_args = NULL,
  bias_func = matrix(0, n, p),
  bias_func_args = NULL
)
}
\arguments{
\item{n}{The number of samples.}

\item{p}{The number of features.}

\item{base}{The base dataframe to use as basis for the simulated data.}

\item{beta}{A (q*p) matrix to denote the effect of each feature}

\item{noise_func}{A function taking the name of the noise function
(e.g. random_noise, uniform_noise)}

\item{noise_func_args}{Additional arguments to be passed to \code{noise_func}}

\item{bias_func}{A function taking the name of the bias function
(e.g. constant_batch_effect)}

\item{bias_func_args}{Additional arguments to be passed to \code{bias_func}}
}
\value{
A data frame of size n*p with simulated data
}
\description{
\code{simulate_dataset()} returns a simulated dataset.
}
\details{
see Readme.md
}
\examples{
\dontrun{
simulate_dataset(n = 5, p = 5, base = data.frame(matrix(rnorm(6 * 6), 6, 6)))
}
\dontrun{
simulate_dataset(10, 15)
}
\dontrun{
simulate_dataset(
  n = 10,
  p = 10,
  noise_func = poisson_noise,
  noise_func_args = list(lambda = 1)
)
}
\dontrun{
simulate_dataset(
  n = 10,
  p = 10,
  noise_func = uniform_noise,
  noise_func_args = list(min = 0, max = 1),
  bias_func = constant_batch_effect,
  bias_func_args = list(
    b = c(1, 1, 1, 2, 2, 2, 3, 3, 4, 4),
    f = 4,
    s = c(0, 1, 2, 3)
  )
)
}
\dontrun{
simulate_dataset(
  n = 10,
  p = 10,
  beta = matrix(1, 1, 10),
  noise_func = uniform_noise,
  noise_func_args = list(min = 0, max = 1),
  bias_func = constant_batch_effect,
  bias_func_args = list(
    b = 3,
    f = 4,
    s = c(2, 3)
  )
)
}

}
