% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_dataset.R
\name{simulate_dataset}
\alias{simulate_dataset}
\title{Simulate a Dataset}
\arguments{
\item{n}{The number of samples.}

\item{p}{The number of features.}

\item{beta}{The model coefficients.}

\item{base}{The dataset to use as basis for the generation, e.g. HG-U133B for
mRNA datasets measured with the Affymetrix HG-U133B microarray.}

\item{platform}{The platform for which data should be simulated.}

\item{family}{The type of model to use when generating the outcome variable.
Either \code{gaussian},\code{binomial} or \code{cox}.}

\item{cor}{Correlation strength within the data}

\item{cortype}{Method to use for generating the correlation within
the dataset.}

\item{noise_func}{A function taking the matrix as
first argument and other optional parameters as futher arguments}

\item{noise_func_args}{Additional arguments to be passed to \code{noise_func}}

\item{noise}{Amount of noise to add to each predictor}

\item{noisevar}{Variance of noise}

\item{bias}{Amount of bias to simulate}

\item{biastype}{Method to use for generating the bias. See below for details.}
}
\value{
A matrix of size n*p.
}
\description{
\code{simulate_dataset()} returns a dataset of size n*p
}
\details{
simulate_dataset <-
function( n=3, p=4, beta=NULL, base="HG-U133B", family="gaussian", cor=0.5,
cortype=1, noise=1, noisevar=1, bias=0, biastype=1 )
}
\examples{
\dontrun{
x <- simulate_dataset(
  n = 100, p = 1000, base =
    "HG-U133B"
)
# should return a matrix of size 100*1000 with genes measured by the
# HG-U133B microarray as features
x2 <- simulate(n, p,
  gse = matrix(rnorm(25), 5, 5), noise_func = unif_noise,
  noise_func_args = list(min = 0, max = 1)
)
}
}
