% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatr.R
\docType{package}
\name{simulatr}
\alias{simulatr}
\title{simulatr: A package that powers portable and scalable simulation studies}
\description{
Simulation studies play a crucial role in the development and assessment of statistical methods. However, writing portable and scalable simulation code can be challenging. Simulatr decouples the \emph{specification} of a simulation from the \emph{execution} of a simulation, enabling users to develop a simulation study on their laptop and seamlessly run the simulation at-scale on a distributed computing platform. Supported platforms include high-performance computing clusters and cloud computing services (e.g., Microsoft Azure, Amazon Web Services, etc.).
}
