% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_constructors.R
\name{simulatr_function}
\alias{simulatr_function}
\title{simulatr_function}
\usage{
simulatr_function(
  f,
  arg_names = NA_character_,
  packages = NA_character_,
  loop = FALSE,
  one_rep_time = NA_real_,
  mult_time_factor = 1.5,
  add_time_factor = 30
)
}
\arguments{
\item{f}{a function with one or more arguments.}

\item{arg_names}{the (ordered) names of parameters (from the parameter grid or list of fixed parameters) to pass to f.}

\item{packages}{packages to load before calling f, stored as a character vector.}

\item{loop}{boolean indicating whether f should be called within a loop.}
}
\value{
A \code{simulatr_function} object
}
\description{
A \code{simulator_function} is a class that stores a method or data generation function alongside associated metadata.
}
