% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_classes.R
\name{simulatr_specifier}
\alias{simulatr_specifier}
\title{simulatr_specifier}
\usage{
simulatr_specifier(
  parameter_grid,
  fixed_parameters,
  generate_data_function,
  run_method_functions
)
}
\arguments{
\item{parameter_grid}{a named data frame storing the values of study parameters that vary}

\item{fixed_parameters}{a named list storing the values of the parameters that stay fixed across parameter settings}

\item{generate_data_function}{a \code{simulatr_function} that generates the data}

\item{run_method_functions}{a named list of \code{simulatr_functions} that define the methods to be run in the simulation}
}
\value{
a \code{simulatr_specifier} object
}
\description{
A \code{simulatr_specifier} object defines a simulation study in the \code{simulatr} framework.
}
\examples{
# an example data generation function
generate_lm_data <- function(beta_0, beta_1, sigma, n) {
x <- rnorm(n)
ep <- rnorm(n, sd = sigma)
y <- beta_0 + beta_1 * x + ep
data.frame(y = y, x = x)
}
data_generator <- simulatr_function(
f = generate_lm_data,
arg_names = c("beta_0", "beta_1", "sigma", "n"),
loop = TRUE)

# an example method function
fit_lm <- function(df) {
fit <- lm(y ~ x, data = df)
s <- summary(fit)$coefficients
row.names(s) <- NULL
out <- data.frame(parameter = c("beta_0", "beta_1"),
Estimate = s[,"Estimate"],
p_val = s[,"Pr(>|t|)"]) \%>\% pivot_longer(
cols = c("Estimate", "p_val"),
names_to = "target")
return(out)
}
method <- simulatr_function(f = fit_lm, packages = "tidyr")

# defining the additional parameters to pass to simulatr_specifier
parameter_grid <- expand.grid(beta_0 = seq(-2, 2, 1), beta_1 = seq(-2, 2, 1))
fixed_parameters <- list(n = 1000, sigma = 1)

# finally, obtain simulatr_specifier object
simulation <- simulatr_specifier(
parameter_grid = parameter_grid,
fixed_parameters = fixed_parameters,
generate_data_function = data_generator,
run_method_functions = list(lm = generate_lm_data))
}
