% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify.model.R
\name{specify.model}
\alias{specify.model}
\title{Specify model}
\usage{
specify.model(type, variables, expression, name.out = "new", export = F)
}
\arguments{
\item{type}{Either "binary" or "continuous".}

\item{variables}{A character vector containing the names of the population and stage variables included in the model or NA (where there are no variables involved).}

\item{expression}{A character vector specifying the construction of the model as a single string.}
}
\value{
a model object
}
\description{
Creates a model object that specifies the details of a model to be used in simulation. Only produces level-1 models (i.e. no models can be included as variables)
}
\examples{
# Run the function
model <- specify.model(type = "binary", variables = NA, expression = "runif(1) > 0.5")
}
